refactor(Running): Hit, Ready 상태 Running으로 추상화
